requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==0.16.1
